[REFACTOR] constraint 관련 수정 및 mainActivity Package 경로 수정
