+++ .idea/misc.xml
@@ -9,6 +9,7 @@
         <entry key="app/src/main/res/layout/activity_search.xml" value="0.19322916666666667" />
         <entry key="app/src/main/res/layout/fragment_issue.xml" value="0.19322916666666667" />
         <entry key="app/src/main/res/layout/fragment_notification.xml" value="0.19322916666666667" />
+        <entry key="app/src/main/res/menu/home_toolbar_menu.xml" value="0.33177083333333335" />
       </map>
     </option>
   </component>
